+++ disco/src/main/java/com/github/lawben/disco/Event.java
@@ -3,16 +3,16 @@ package com.github.lawben.disco;
 public class Event {
     public static final int NO_KEY = -2;
 
-    private final Integer value;
+    private final long value;
     private final long timestamp;
     private final int streamId;
     private final int key;
 
-    public Event(Integer value, long timestamp, int streamId) {
+    public Event(long value, long timestamp, int streamId) {
         this(value, timestamp, streamId, NO_KEY);
     }
 
-    public Event(Integer value, long timestamp, int streamId, int key) {
+    public Event(long value, long timestamp, int streamId, int key) {
         this.value = value;
         this.timestamp = timestamp;
         this.streamId = streamId;
@@ -23,12 +23,12 @@ public class Event {
         final String[] eventParts = eventString.split(",");
         final int streamId = Integer.parseInt(eventParts[0]);
         final long eventTimestamp = Long.parseLong(eventParts[1]);
-        final int eventValue = Integer.parseInt(eventParts[2]);
+        final long eventValue = Long.parseLong(eventParts[2]);
         final int key = eventParts.length == 4 ? Integer.parseInt(eventParts[3]) : NO_KEY;
         return new Event(eventValue, eventTimestamp, streamId, key);
     }
 
-    public Integer getValue() {
+    public long getValue() {
         return value;
     }
